Add an HTTP server to handle Prometheus scrapes
Prometheus sends HTTP GET requests to an application at `/metrics` endpoint
to scrape stats from it. This commits adds a listener for these requests,
and a handler to response with stats to Prometheus.

The listener needs to be instantiated by the application, with a set of
callbacks, via the `CollectorRegistry` module, that can be called by the
handler to prepare the response messages.